Update ICU to 76.1 commit_hash:908cd3760bb1e3e7ffecef62acf9b632d8ca31a4

+++ contrib/libs/icu/io/sscanf.cpp
@@ -73,7 +73,7 @@ u_vsscanf(const char16_t   *buffer,
     int32_t size = (int32_t)uprv_strlen(patternSpecification) + 1;
 
     /* convert from the default codepage to Unicode */
-    if (size >= (int32_t)MAX_UCHAR_BUFFER_SIZE(patBuffer)) {
+    if (size >= MAX_UCHAR_BUFFER_SIZE(patBuffer)) {
         pattern = (char16_t *)uprv_malloc(size * sizeof(char16_t));
         if (pattern == nullptr) {
             return 0;
